[ASTPrinter] Exclude attributes in parameter flags
The AST printer uses `PrintOptions` to determine what parts of the AST
to print. The `PrintOptions.SkipAttributes` option can be used to skip
printing declaration and type attributes. This flag was not being
respected when printing parameter flags, leading to attributes like
`@escaping` and `@Sendable` on parameter flags being printed even when
the skip option was enabled. This patch updates the parameter flag
processing to short-circuit appropriately if this option is enabled.

It also appropriately removes the now redundant logic for excluding
these attributes in subheadings manually in the symbol graph generator.
The configuration that excludes attributes in the declaration fragments
of symbols performed a number of string operations through a `String ->
AnyAttrKind` map to first include all attributes in the exclusion list,
and selectively remove the ones that are required. This logic has been
made more readable and less computationally expensive by using a static
array to record the attributes that are required, and exclude them at
insertion time.